+++ app/routes/pokemon.js
@@ -1,0 +1,14 @@
+const express = require('express');
+const router = express.Router();
+
+router.get('/', (req,res) => {
+    res.status(200).send("Dona'm la id d'un pokemon per buscar-ne les dades a la Pokedex.")
+})
+
+//https://github.com/everydeveloper/node-express-course/blob/master/responses/04-get-data-with-var.md
+router.get('/:id', (req,res) => {
+    const pokeID = req.params.id;
+    res.send("Buscaré el Pokemon amb ID "+pokeID)
+})
+
+module.exports = router;
